户型分析建筑面积（类型：int->decimal）

+++ src/main/java/com/jiubo/buildstore/bean/BuildingAnalysisBean.java
@@ -124,7 +124,7 @@ public class BuildingAnalysisBean implements Serializable {
     /**
      * 建筑面积
      */
-    private Integer buildArea;
+    private BigDecimal buildArea;
 
     /**
      * 居室
